update(heredoc): sep files
