+++ includes/ms_parse.h
@@ -88,7 +88,9 @@ t_result	execute_heredoc_each(t_redirect *redirect, bool is_interactive);
 void		move_redirect_from_command(t_ast *ast_node);
 char		*create_heredoc_filename(void);
 t_result	create_filename_and_open_heredoc_fd(int *fd, char **filename);
-
+t_result	read_input_save_to_fd(int fd, \
+									const char *delimiter, \
+									bool is_interactive);
 /* is */
 bool		is_node_kind_subshell(t_node_kind node_kind);
 bool		is_node_kind_exec_heredoc(t_node_kind node_kind);
